refactor: separate floating button from scroll-to-top button for modularity
